refactor: rename extension directory
