+++ _emulator/extensions/firestore-record-user-acknowledgements.env.local
