--- firestore-notice-extension/functions/__tests__/jest.setup.ts
@@ -1,12 +1,0 @@
-import path from "path";
-
-(async function () {
-  require("dotenv").config({
-    path: path.resolve(
-      __dirname,
-      "../../../_emulator/extensions/firestore-notice-extension.env.local"
-    ),
-  });
-
-  process.env.EXT_INSTANCE_ID = "firestore-notice-extension";
-})();
