+++ firestore-record-user-acknowledgements/admin-dashboard/LICENSE
