+++ firestore-record-user-acknowledgements/admin-dashboard/app/components/Label.tsx
