+++ firestore-record-user-acknowledgements/admin-dashboard/app/firebase.server.ts
@@ -1,14 +1,16 @@
-import { initializeApp, getApps, applicationDefault } from 'firebase-admin/app';
-import { getFirestore, FieldValue } from 'firebase-admin/firestore';
-import type { QuerySnapshot } from 'firebase-admin/firestore';
+import { initializeApp, getApps, applicationDefault } from "firebase-admin/app";
+import { getFirestore, FieldValue } from "firebase-admin/firestore";
+import type { QuerySnapshot } from "firebase-admin/firestore";
 import type { Acknowledgement, Notice } from "./types";
 
 const projectId = process.env.PROJECT_ID;
-const noticesCollectionPath = process.env.NOTICES_COLLECTION_PATH || 'notices';
+const noticesCollectionPath = process.env.NOTICES_COLLECTION_PATH || "notices";
 
 if (getApps().length === 0) {
   if (!projectId) {
-    throw new Error('Missing PROJECT_ID environment variable. Please provide a .env file with the PROJECT_ID variable.');
+    throw new Error(
+      "Missing PROJECT_ID environment variable. Please provide a .env file with the PROJECT_ID variable."
+    );
   }
 
   initializeApp({
@@ -20,7 +22,9 @@ if (getApps().length === 0) {
 const firestore = getFirestore();
 
 // Converts a Firestore document to a plain object.
-function documentToObject<T>(document: FirebaseFirestore.DocumentSnapshot<FirebaseFirestore.DocumentData>): T {
+function documentToObject<T>(
+  document: FirebaseFirestore.DocumentSnapshot<FirebaseFirestore.DocumentData>
+): T {
   return {
     ...document.data(),
     id: document.id,
@@ -29,7 +33,7 @@ function documentToObject<T>(document: FirebaseFirestore.DocumentSnapshot<Fireba
 
 // Converts a Firestore document to an array.
 function snapshotToArray<T>(snapshot: QuerySnapshot): T[] {
-  return snapshot.docs.map(doc => documentToObject<T>(doc));
+  return snapshot.docs.map((doc) => documentToObject<T>(doc));
 }
 
 // Returns a list of notices.
@@ -39,8 +43,11 @@ export async function getNotices(): Promise<Notice[]> {
 }
 
 // Returns a notice, if it exists.
-export async function getNotice(id: string): Promise<Notice | null>  {
-  const snapshot = await firestore.collection(noticesCollectionPath).doc(id).get();
+export async function getNotice(id: string): Promise<Notice | null> {
+  const snapshot = await firestore
+    .collection(noticesCollectionPath)
+    .doc(id)
+    .get();
 
   if (snapshot.exists) {
     return documentToObject<Notice>(snapshot);
@@ -50,15 +57,17 @@ export async function getNotice(id: string): Promise<Notice | null>  {
 }
 
 // Returns a list of acknowledgements.
-export async function getAcknowledgements(noticeId: string): Promise<Acknowledgement[]> {
+export async function getAcknowledgements(
+  noticeId: string
+): Promise<Acknowledgement[]> {
   const snapshot = await firestore
     .collection(noticesCollectionPath)
     .doc(noticeId)
     .collection("acknowledgements")
-    .orderBy('createdAt', 'desc')
+    .orderBy("createdAt", "desc")
     .get();
 
-    return snapshotToArray<Acknowledgement>(snapshot);
+  return snapshotToArray<Acknowledgement>(snapshot);
 }
 
 // Creates a new notice, and returns the new ID.
