+++ firestore-record-user-acknowledgements/admin-dashboard/app/types.ts
