+++ firestore-record-user-acknowledgements/admin-dashboard/package.json
