+++ firestore-record-user-acknowledgements/admin-dashboard/tailwind.config.js
@@ -5,4 +5,4 @@ module.exports = {
     extend: {},
   },
   plugins: [],
-}
+};
