+++ firestore-record-user-acknowledgements/admin-dashboard/tsconfig.json
