+++ firestore-record-user-acknowledgements/functions/.gitignore
