+++ firestore-record-user-acknowledgements/functions/__tests__/functions.test.ts
