+++ firestore-record-user-acknowledgements/functions/__tests__/helpers/setupEnvironment.ts
