+++ firestore-record-user-acknowledgements/functions/__tests__/jest.setup.ts
@@ -1,0 +1,12 @@
+import path from "path";
+
+(async function () {
+  require("dotenv").config({
+    path: path.resolve(
+      __dirname,
+      "../../../_emulator/extensions/firestore-record-user-acknowledgements.env.local"
+    ),
+  });
+
+  process.env.EXT_INSTANCE_ID = "firestore-record-user-acknowledgements";
+})();
