+++ firestore-record-user-acknowledgements/functions/__tests__/tsconfig.json
