+++ firestore-record-user-acknowledgements/functions/jest.config.js
