+++ firestore-record-user-acknowledgements/functions/src/config.ts
