+++ firestore-record-user-acknowledgements/functions/src/index.ts
