+++ firestore-record-user-acknowledgements/functions/src/interface.ts
