+++ firestore-record-user-acknowledgements/functions/src/logs.ts
