+++ firestore-record-user-acknowledgements/functions/tsconfig.json
@@ -2,8 +2,7 @@
   "compilerOptions": {
     "target": "ES2020",
     "lib": ["ES2020", "DOM"],
-    "types": ["node", "jest", "@types/jest"],
-    "typeRoots": ["./node_modules/@types"],
+    "types": ["jest", "node"],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": false,
