+++ firestore-record-user-acknowledgements/tsconfig.json
